+++ lib/release_packer.dart
@@ -1,0 +1,4 @@
+/// Release Updater Packer Library.
+library release_updater.packer;
+
+export 'src/release_packer.dart';