+++ lib/release_updater.dart
@@ -2,7 +2,8 @@
 library release_updater;
 
 export 'src/release_updater_base.dart';
+export 'src/release_updater_bundle.dart';
+export 'src/release_updater_platform.dart';
 export 'src/release_updater_provider.dart';
 export 'src/release_updater_provider_http.dart';
-export 'src/release_updater_release_bundle.dart';
 export 'src/release_updater_storage.dart';